Use mysql as the database
